hyperledger-fabric-kuma-token-example
This a simple example of a token created on the Hyperledger Fabric platform. It exists out of a simple wallet system that can be used to exchange Kuma-tokens and a multisig system that can be used for setting up multisig contracts between users. This example shows how the different tools we created for the Hyperledger Fabric platform can be used: